refactor into happy code
